34 location openinghours merge

+++ server/src/main/java/com/codecool/repository/LocationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface LocationRepository extends JpaRepository<Location, Long> {
-    int deleteLocationById(long id);
+    long deleteLocationById(long id);
 }
